+++ prawcore/auth.py
@@ -6,32 +6,26 @@ from requests import Request
 from requests.status_codes import codes
 
 
-class Authenticator(object):
-    """Stores OAuth2 authentication credentials."""
+class BaseAuthenticator(object):
+    """Provide the base authenticator object that stores OAuth2 credentials."""
 
-    def __init__(self, requestor, client_id, client_secret=None,
-                 redirect_uri=None):
-        """Represent a single authentication to reddit's API.
+    def __init__(self, requestor, client_id, redirect_uri=None):
+        """Represent a single authentication to Reddit's API.
 
         :param requestor: An instance of :class:`Requestor`.
         :param client_id: The OAuth2 client ID to use with the session.
-        :param client_secret: (optional) The OAuth2 client secret to use with
-            the session. This parameter is required for use with all
-            Authorizers save for DeviceIDAuthorizer.
         :param redirect_uri: (optional) The redirect URI exactly as specified
-            in your OAuth application settings on reddit. This parameter is
+            in your OAuth application settings on Reddit. This parameter is
             required if you want to use the ``authorize_url`` method, or the
             ``authorize`` method of the ``Authorizer`` class.
 
         """
         self._requestor = requestor
         self.client_id = client_id
-        self.client_secret = client_secret
         self.redirect_uri = redirect_uri
 
     def _post(self, url, success_status=codes['ok'], **data):
-        auth = (self.client_id, self.client_secret or '')
-        response = self._requestor.request('post', url, auth=auth,
+        response = self._requestor.request('post', url, auth=self._auth(),
                                            data=sorted(data.items()))
         if response.status_code != success_status:
             raise ResponseException(response)
@@ -55,17 +49,18 @@ class Authenticator(object):
             raise InvalidInvocation('redirect URI not provided')
 
         params = {'client_id': self.client_id, 'duration': duration,
-                  'redirect_uri': self.redirect_uri, 'response_type': 'code',
+                  'redirect_uri': self.redirect_uri,
+                  'response_type': self.RESPONSE_TYPE,
                   'scope': ' '.join(scopes), 'state': state}
         url = self._requestor.reddit_url + const.AUTHORIZATION_PATH
         request = Request('GET', url, params=params)
         return request.prepare().url
 
     def revoke_token(self, token, token_type=None):
-        """Ask reddit to revoke the provided token.
+        """Ask Reddit to revoke the provided token.
 
         :param token: The access or refresh token to revoke.
-        :param token_type: (Optional) When provided, hint to reddit what the
+        :param token_type: (Optional) When provided, hint to Reddit what the
             token type is for a possible efficiency gain. The value can be
             either ``access_token`` or ``refresh_token``.
 
@@ -77,19 +72,71 @@ class Authenticator(object):
         self._post(url, success_status=codes['no_content'], **data)
 
 
-class Authorizer(object):
-    """Manages OAuth2 authorization tokens and scopes."""
+class TrustedAuthenticator(BaseAuthenticator):
+    """Store OAuth2 authentication credentials for web, or script type apps."""
 
-    def __init__(self, authenticator, refresh_token=None):
-        """Represent a single authorization to reddit's API.
+    RESPONSE_TYPE = 'code'
 
-        :param authenticator: An instance of :class:`Authenticator`.
-        :param refresh_token: (Optional) Enables the ability to refresh the
-            authorization.
+    def __init__(self, requestor, client_id, client_secret, redirect_uri=None):
+        """Represent a single authentication to Reddit's API.
+
+        :param requestor: An instance of :class:`Requestor`.
+        :param client_id: The OAuth2 client ID to use with the session.
+        :param client_secret: The OAuth2 client secret to use with the session.
+        :param redirect_uri: (optional) The redirect URI exactly as specified
+            in your OAuth application settings on Reddit. This parameter is
+            required if you want to use the ``authorize_url`` method, or the
+            ``authorize`` method of the ``Authorizer`` class.
+
+        """
+        super(TrustedAuthenticator, self).__init__(requestor, client_id,
+                                                   redirect_uri)
+        self.client_secret = client_secret
+
+    def _auth(self):
+        return (self.client_id, self.client_secret)
+
+
+class UntrustedAuthenticator(BaseAuthenticator):
+    """Store OAuth2 authentication credentials for installed applications."""
+
+    RESPONSE_TYPE = 'token'
+
+    def _auth(self):
+        return (self.client_id, '')
+
+    def authorize_url(self, scopes, state):
+        """Return the URL used out-of-band to grant access to your application.
+
+        :param scopes: A list of OAuth scopes to request authorization for.
+        :param state: A string that will be reflected in the callback to
+            ``redirect_uri``. This value should be temporarily unique to the
+            client for whom the URL was generated for.
+
+        """
+        return super(UntrustedAuthenticator, self).authorize_url(
+            'temporary', scopes, state)
+
+    def revoke_token(self, token):
+        """Ask Reddit to revoke the provided token.
+
+        :param token: The access token to revoke.
+
+        """
+        return super(UntrustedAuthenticator, self).revoke_token(token,
+                                                                'access_token')
+
+
+class BaseAuthorizer(object):
+    """Superclass for OAuth2 authorization tokens and scopes."""
+
+    def __init__(self, authenticator):
+        """Represent a single authorization to Reddit's API.
+
+        :param authenticator: An instance of :class:`BaseAuthenticator`.
 
         """
         self._authenticator = authenticator
-        self.refresh_token = refresh_token
         self._clear_access_token()
         self._validate_authenticator()
 
@@ -116,21 +163,9 @@ class Authorizer(object):
         self.scopes = set(payload['scope'].split(' '))
 
     def _validate_authenticator(self):
-        if not self._authenticator.client_secret:
-            raise InvalidInvocation('Authorizer must be used with an '
-                                    'Authenticator that has a client secret.')
-
-    def authorize(self, code):
-        """Obtain and set authorization tokens based on ``code``.
-
-        :param code: The code obtained by an out-of-band authorization request
-            to reddit.
-
-        """
-        if self._authenticator.redirect_uri is None:
-            raise InvalidInvocation('redirect URI not provided')
-        self._request_token(code=code, grant_type='authorization_code',
-                            redirect_uri=self._authenticator.redirect_uri)
+        if not isinstance(self._authenticator, self.AUTHENTICATOR_CLASS):
+            raise InvalidInvocation('Must use a authenticator of type {}.'
+                                    .format(self.AUTHENTICATOR_CLASS.__name__))
 
     def is_valid(self):
         """Return whether or not the Authorizer is ready to authorize requests.
@@ -141,6 +176,43 @@ class Authorizer(object):
         """
         return self.access_token is not None \
             and time.time() < self._expiration_timestamp
+
+    def revoke(self):
+        """Revoke the current Authorization."""
+        if self.access_token is None:
+            raise InvalidInvocation('no token available to revoke')
+
+        self._authenticator.revoke_token(self.access_token, 'access_token')
+        self._clear_access_token()
+
+
+class Authorizer(BaseAuthorizer):
+    """Manages OAuth2 authorization tokens and scopes."""
+
+    AUTHENTICATOR_CLASS = TrustedAuthenticator
+
+    def __init__(self, authenticator, refresh_token=None):
+        """Represent a single authorization to Reddit's API.
+
+        :param authenticator: An instance of :class:`TrustedAuthenticator`.
+        :param refresh_token: (Optional) Enables the ability to refresh the
+            authorization.
+
+        """
+        super(Authorizer, self).__init__(authenticator)
+        self.refresh_token = refresh_token
+
+    def authorize(self, code):
+        """Obtain and set authorization tokens based on ``code``.
+
+        :param code: The code obtained by an out-of-band authorization request
+            to Reddit.
+
+        """
+        if self._authenticator.redirect_uri is None:
+            raise InvalidInvocation('redirect URI not provided')
+        self._request_token(code=code, grant_type='authorization_code',
+                            redirect_uri=self._authenticator.redirect_uri)
 
     def refresh(self):
         """Obtain a new access token from the refresh_token."""
@@ -159,30 +231,29 @@ class Authorizer(object):
         associated with that authorization.
 
         """
-        if self.refresh_token is not None and not only_access:
+        if only_access or self.refresh_token is None:
+            super(Authorizer, self).revoke()
+        else:
             self._authenticator.revoke_token(self.refresh_token,
                                              'refresh_token')
-        elif self.access_token is not None:
-            self._authenticator.revoke_token(self.access_token, 'access_token')
-        else:
-            raise InvalidInvocation('no token available to revoke')
-
-        self._clear_access_token()
-        if not only_access:
+            self._clear_access_token()
             self.refresh_token = None
 
 
-class DeviceIDAuthorizer(Authorizer):
+class DeviceIDAuthorizer(BaseAuthorizer):
     """Manages app-only OAuth2 for 'installed' applications.
 
     While the '*' scope will be available, some endpoints simply will not work
-    due to the lack of an associated reddit account.
+    due to the lack of an associated Reddit account.
+
     """
+
+    AUTHENTICATOR_CLASS = UntrustedAuthenticator
 
     def __init__(self, authenticator, device_id='DO_NOT_TRACK_THIS_DEVICE'):
         """Represent an app-only OAuth2 authorization for 'installed' apps.
 
-        :param authenticator: An instance of :class:`Authenticator`.
+        :param authenticator: An instance of :class:`UntrustedAuthenticator`.
         :param device_id: (optional) A unique ID (20-30 character ASCII string)
             (default DO_NOT_TRACK_THIS_DEVICE). For more information about this
             parameter, see:
@@ -192,9 +263,8 @@ class DeviceIDAuthorizer(Authorizer):
         self._device_id = device_id
 
     def _validate_authenticator(self):
-        if self._authenticator.client_secret:
-            raise InvalidInvocation('DeviceIDAuthorizor cannot be used with an'
-                                    ' Authenticator that has a client secret.')
+        if not isinstance(self._authenticator, UntrustedAuthenticator):
+            raise InvalidInvocation('Must use an UntrustedAuthentictor.')
 
     def refresh(self):
         """Obtain a new access token."""
@@ -203,11 +273,39 @@ class DeviceIDAuthorizer(Authorizer):
                             device_id=self._device_id)
 
 
+class ImplicitAuthorizer(BaseAuthorizer):
+    """Manages implicit installed-app type authorizations."""
+
+    AUTHENTICATOR_CLASS = UntrustedAuthenticator
+
+    def __init__(self, authenticator, access_token, expires_in, scope):
+        """Represent a single implicit authorization to Reddit's API.
+
+        :param authenticator: An instance of :class:`UntrustedAuthenticator`.
+        :param access_token: The access_token obtained from Reddit via callback
+            to the authenticator's redirect_uri.
+        :param expires_in: The number of seconds the ``access_token`` is valid
+            for. The origin of this value was returned from Reddit via callback
+            to the authenticator's redirect uri. Note, you may need to subtract
+            an offset before passing in this number to account for a delay
+            between when Reddit prepared the response, and when you make this
+            function call.
+        :param scopes: A space-delimited string of Reddit OAuth2 scope names as
+            returned from Reddit in the callback to the authenticator's
+            redirect uri.
+
+        """
+        super(ImplicitAuthorizer, self).__init__(authenticator)
+        self._expiration_timestamp = time.time() + expires_in
+        self.access_token = access_token
+        self.scopes = set(scope.split(' '))
+
+
 class ReadOnlyAuthorizer(Authorizer):
-    """Manages authorizations that are not associated with a reddit account.
+    """Manages authorizations that are not associated with a Reddit account.
 
     While the '*' scope will be available, some endpoints simply will not work
-    due to the lack of an associated reddit account.
+    due to the lack of an associated Reddit account.
 
     """
 
@@ -224,11 +322,13 @@ class ScriptAuthorizer(Authorizer):
 
     """
 
-    def __init__(self, authenticator, username, password):
-        """Represent a single personal-use authorization to reddit's API.
+    AUTHENTICATOR_CLASS = TrustedAuthenticator
 
-        :param authenticator: An instance of :class:`Authenticator`.
-        :param username: The reddit username of one of the application's
+    def __init__(self, authenticator, username, password):
+        """Represent a single personal-use authorization to Reddit's API.
+
+        :param authenticator: An instance of :class:`TrustedAuthenticator`.
+        :param username: The Reddit username of one of the application's
             developers.
         :param password: The password associated with ``username``.